New upstream version 202303.3

+++ rdkit/ML/ScreenComposite.py
@@ -109,8 +109,8 @@ a file containing a pickled composite model and _filename_ is a QDAT file.
 
 """
 
-
 import os
+import pickle
 import sys
 
 import numpy
@@ -120,8 +120,6 @@ from rdkit.Dbase import DbModule
 from rdkit.Dbase.DbConnection import DbConnect
 from rdkit.ML import CompositeRun
 from rdkit.ML.Data import DataUtils, SplitData
-import pickle
-
 
 try:
   from PIL import Image, ImageDraw
@@ -426,7 +424,7 @@ def ShowVoteResults(indices, data, composite, nResultCodes, threshold, verbose=1
     print('average correct confidence:   % 6.4f' % avgGood)
     print('average incorrect confidence: % 6.4f' % avgBad)
 
-  voteTab = numpy.zeros((nResultCodes, nResultCodes), numpy.int)
+  voteTab = numpy.zeros((nResultCodes, nResultCodes), numpy.int32)
   for res in goodRes:
     voteTab[res, res] += 1
   for ans, res, conf, idx in badVotes:
@@ -719,12 +717,12 @@ def ScreenFromDetails(models, details, callback=None, setup=None, appendExamples
   if setup is not None:
     setup(nModels * data.GetNPts())
 
-  nGood = numpy.zeros(nModels, numpy.float)
-  nBad = numpy.zeros(nModels, numpy.float)
-  nSkip = numpy.zeros(nModels, numpy.float)
-  confGood = numpy.zeros(nModels, numpy.float)
-  confBad = numpy.zeros(nModels, numpy.float)
-  confSkip = numpy.zeros(nModels, numpy.float)
+  nGood = numpy.zeros(nModels, float)
+  nBad = numpy.zeros(nModels, float)
+  nSkip = numpy.zeros(nModels, float)
+  confGood = numpy.zeros(nModels, float)
+  confBad = numpy.zeros(nModels, float)
+  confSkip = numpy.zeros(nModels, float)
   voteTab = None
   if goodVotes is None:
     goodVotes = []
@@ -787,12 +785,13 @@ def ScreenFromDetails(models, details, callback=None, setup=None, appendExamples
       errorEstimate = 0
     else:
       errorEstimate = 1
-    g, b, s, aG, aB, aS, vT = ShowVoteResults(
-      testIdx, data, model, nPossible[-1], details.threshold, verbose=0, callback=cb,
-      appendExamples=appendExamples, goodVotes=goodVotes, badVotes=badVotes, noVotes=noVotes,
-      errorEstimate=errorEstimate)
+    g, b, s, aG, aB, aS, vT = ShowVoteResults(testIdx, data, model, nPossible[-1],
+                                              details.threshold, verbose=0, callback=cb,
+                                              appendExamples=appendExamples, goodVotes=goodVotes,
+                                              badVotes=badVotes, noVotes=noVotes,
+                                              errorEstimate=errorEstimate)
     if voteTab is None:
-      voteTab = numpy.zeros(vT.shape, numpy.float)
+      voteTab = numpy.zeros(vT.shape, float)
     if hasattr(details, 'errorAnalysis') and details.errorAnalysis:
       for a, p, c, idx in badVotes:
         label = testIdx[idx]
@@ -845,8 +844,8 @@ def ScreenFromDetails(models, details, callback=None, setup=None, appendExamples
     avgConfSkip = sum(confSkip) / nModels
     devConfSkip = numpy.sqrt(sum((confSkip - avgConfSkip)**2) / (nModels - 1))
     return ((avgNGood, devNGood), (avgNBad, devNBad), (avgNSkip, devNSkip),
-            (avgConfGood, devConfGood), (avgConfBad, devConfBad), (avgConfSkip, devConfSkip),
-            voteTab)
+            (avgConfGood, devConfGood), (avgConfBad, devConfBad), (avgConfSkip,
+                                                                   devConfSkip), voteTab)
 
 
 def GetScreenImage(nGood, nBad, nRej, size=None):
@@ -944,8 +943,8 @@ def ScreenToHtml(nGood, nBad, nRej, avgGood, avgBad, avgSkip, voteTable, imgDir=
   nPoss = len(voteTable)
   pureCounts = numpy.sum(voteTable, 1)
   accCounts = numpy.sum(voteTable, 0)
-  pureVect = numpy.zeros(nPoss, numpy.float)
-  accVect = numpy.zeros(nPoss, numpy.float)
+  pureVect = numpy.zeros(nPoss, float)
+  accVect = numpy.zeros(nPoss, float)
   for i in range(nPoss):
     if pureCounts[i]:
       pureVect[i] = float(voteTable[i, i]) / pureCounts[i]
@@ -1219,14 +1218,16 @@ def ShowVersion(includeArgs=0):
 def ParseArgs(details):
   import getopt
   try:
-    args, extras = getopt.getopt(sys.argv[1:], 'EDd:t:VN:HThSRF:v:AX', ['predPlot=',
-                                                                        'predActCol=',
-                                                                        'predActTable=',
-                                                                        'predLogScale',
-                                                                        'predShow',
-                                                                        'OOB',
-                                                                        'pickleCol=',
-                                                                        'enrich=', ])
+    args, extras = getopt.getopt(sys.argv[1:], 'EDd:t:VN:HThSRF:v:AX', [
+      'predPlot=',
+      'predActCol=',
+      'predActTable=',
+      'predLogScale',
+      'predShow',
+      'OOB',
+      'pickleCol=',
+      'enrich=',
+    ])
   except Exception:
     import traceback
     traceback.print_exc()
@@ -1397,14 +1398,14 @@ if __name__ == '__main__':
       if len(details.screenVoteTol) > 1:
         message('\n-----*****-----*****-----*****-----*****-----*****-----*****-----\n')
         message('Tolerance: %f' % tol)
-      nGood = numpy.zeros(nModels, numpy.float)
-      nBad = numpy.zeros(nModels, numpy.float)
-      nSkip = numpy.zeros(nModels, numpy.float)
-      confGood = numpy.zeros(nModels, numpy.float)
-      confBad = numpy.zeros(nModels, numpy.float)
-      confSkip = numpy.zeros(nModels, numpy.float)
+      nGood = numpy.zeros(nModels, float)
+      nBad = numpy.zeros(nModels, float)
+      nSkip = numpy.zeros(nModels, float)
+      confGood = numpy.zeros(nModels, float)
+      confBad = numpy.zeros(nModels, float)
+      confSkip = numpy.zeros(nModels, float)
       if details.enrichTgt >= 0:
-        enrichments = numpy.zeros(nModels, numpy.float)
+        enrichments = numpy.zeros(nModels, float)
       goodVoteDict = {}
       badVoteDict = {}
       noVoteDict = {}
@@ -1429,12 +1430,13 @@ if __name__ == '__main__':
             goodVotes = []
           else:
             goodVotes = None
-          g, b, s, aG, aB, aS, vT = ShowVoteResults(
-            testIdx, tmpD, model, nRes, tol, verbose=details.verbose, screenResults=screenRes,
-            badVotes=badVotes, noVotes=noVotes, goodVotes=goodVotes,
-            errorEstimate=details.errorEstimate)
+          g, b, s, aG, aB, aS, vT = ShowVoteResults(testIdx, tmpD, model, nRes, tol,
+                                                    verbose=details.verbose,
+                                                    screenResults=screenRes, badVotes=badVotes,
+                                                    noVotes=noVotes, goodVotes=goodVotes,
+                                                    errorEstimate=details.errorEstimate)
           if voteTab is None:
-            voteTab = numpy.zeros(vT.shape, numpy.float)
+            voteTab = numpy.zeros(vT.shape, float)
           if details.errorAnalysis:
             for a, p, c, idx in badVotes:
               label = testIdx[idx]
@@ -1472,18 +1474,18 @@ if __name__ == '__main__':
           print('id, prediction, confidence, flag(-1=skipped,0=wrong,1=correct)')
           for ans, pred, conf, idx in goodVotes:
             pt = tmpD[testIdx[idx]]
-            assert model.GetActivityQuantBounds() or pt[-1] == ans, 'bad point?: %s != %s' % (
-              str(pt[-1]), str(ans))
+            assert model.GetActivityQuantBounds() or pt[-1] == ans, 'bad point?: %s != %s' % (str(
+              pt[-1]), str(ans))
             print('%s, %d, %.4f, 1' % (str(pt[0]), pred, conf))
           for ans, pred, conf, idx in badVotes:
             pt = tmpD[testIdx[idx]]
-            assert model.GetActivityQuantBounds() or pt[-1] == ans, 'bad point?: %s != %s' % (
-              str(pt[-1]), str(ans))
+            assert model.GetActivityQuantBounds() or pt[-1] == ans, 'bad point?: %s != %s' % (str(
+              pt[-1]), str(ans))
             print('%s, %d, %.4f, 0' % (str(pt[0]), pred, conf))
           for ans, pred, conf, idx in noVotes:
             pt = tmpD[testIdx[idx]]
-            assert model.GetActivityQuantBounds() or pt[-1] == ans, 'bad point?: %s != %s' % (
-              str(pt[-1]), str(ans))
+            assert model.GetActivityQuantBounds() or pt[-1] == ans, 'bad point?: %s != %s' % (str(
+              pt[-1]), str(ans))
             print('%s, %d, %.4f, -1' % (str(pt[0]), pred, conf))
           print('-^-^-^-^-^-^-^-  -^-^-^-^-^-^-^-')
 
@@ -1578,11 +1580,11 @@ if __name__ == '__main__':
       bestSkip = nSkip[bestIdx]
       nClassified = bestGood + bestBad
       nExamples = nClassified + bestSkip
-      print('Misclassifications: \t%%%5.2f   %d / %d' % (100 * bestBad / nExamples, bestBad,
-                                                         nExamples))
+      print('Misclassifications: \t%%%5.2f   %d / %d' %
+            (100 * bestBad / nExamples, bestBad, nExamples))
       if bestSkip > 0:
-        print('\tthreshold: \t%%%5.2f   %d / %d' % (100 * bestBad / nClassified, bestBad,
-                                                    nClassified))
+        print('\tthreshold: \t%%%5.2f   %d / %d' %
+              (100 * bestBad / nClassified, bestBad, nClassified))
         print()
         print('Number Skipped: %%%4.2f    %d' % (100 * bestSkip / nExamples, bestSkip))
